# Conflicts:
#	_maps/map_files/shared/CentCom.dmm
#	code/__DEFINES/jobs.dm
#	code/__DEFINES/storytellers.dm
#	code/__DEFINES/traits.dm
#	code/_globalvars/special_traits/traits.dm
#	code/_globalvars/traits.dm
#	code/datums/character_flaw/_character_flaw.dm
#	code/datums/faith/devotion_subtypes.dm
#	code/datums/gods/faiths/inhumen_pantheon.dm
#	code/datums/gods/patrons/atheism.dm
#	code/datums/gods/patrons/divine_pantheon.dm
#	code/datums/gods/patrons/forgotten.dm
#	code/datums/gods/patrons/inhumen_pantheon.dm
#	code/datums/migrants/waves/crusader_wave.dm
#	code/datums/migrants/waves/daywalker.dm
#	code/datums/mind.dm
#	code/datums/skills/combat.dm
#	code/datums/sleep_advancement.dm
#	code/datums/status_effects/rogue/roguebuff.dm
#	code/game/area/roguetownareas.dm
#	code/game/objects/effects/glowshroom.dm
#	code/game/objects/items/gem_items.dm
#	code/game/objects/items/gems.dm
#	code/game/objects/items/weapons/melee/godweapons.dm
#	code/game/objects/items/weapons/melee/knives.dm
#	code/game/objects/items/weapons/melee/polearms.dm
#	code/game/objects/items/weapons/melee/swords.dm
#	code/game/objects/structures/apiary.dm
#	code/game/objects/structures/fake_machines/mail.dm
#	code/game/objects/structures/hole.dm
#	code/game/turfs/open/floor/roguefloor.dm
#	code/modules/antagonists/villain/assassin.dm
#	code/modules/antagonists/villain/zizo_cult/zizocult.dm
#	code/modules/client/round_end_panel.dm
#	code/modules/client/verbs/ghost.dm
#	code/modules/clothing/neck/psycross.dm
#	code/modules/clothing/ring/misc.dm
#	code/modules/crafting/blueprinting/blueprints/masonry.dm
#	code/modules/events/god_intervention/astrata.dm
#	code/modules/events/god_intervention/graggar.dm
#	code/modules/events/god_intervention/ravox.dm
#	code/modules/events/god_intervention/xylix.dm
#	code/modules/instruments/_instruments.dm
#	code/modules/jobs/job_types/adventurer/types/combat/dredge.dm
#	code/modules/jobs/job_types/adventurer/types/combat/hollowranger.dm
#	code/modules/jobs/job_types/adventurer/types/combat/monk.dm
#	code/modules/jobs/job_types/adventurer/types/combat/rare/swashbuckler.dm
#	code/modules/jobs/job_types/adventurer/types/pilgrim/briar.dm
#	code/modules/jobs/job_types/adventurer/types/pilgrim/hunter.dm
#	code/modules/jobs/job_types/adventurer/types/special/crusader.dm
#	code/modules/jobs/job_types/adventurer/types/wretch/necromancer.dm
#	code/modules/jobs/job_types/adventurer/types/wretch/vigilante.dm
#	code/modules/jobs/job_types/church/monk.dm
#	code/modules/jobs/job_types/church/priest.dm
#	code/modules/jobs/job_types/church/templar.dm
#	code/modules/jobs/job_types/garrison/town_elder.dm
#	code/modules/jobs/job_types/inquistion/absolver.dm
#	code/modules/jobs/job_types/inquistion/adept.dm
#	code/modules/jobs/job_types/inquistion/adept_classes/brutal_zealot.dm
#	code/modules/jobs/job_types/inquistion/inquisitor.dm
#	code/modules/jobs/job_types/nobility/courtphys.dm
#	code/modules/jobs/job_types/nobility/magician.dm
#	code/modules/jobs/job_types/other/deathknight.dm
#	code/modules/jobs/job_types/other/skeleton.dm
#	code/modules/jobs/job_types/peasants/cheesemaker.dm
#	code/modules/jobs/job_types/youngfolk/churchling.dm
#	code/modules/language/otavan.dm
#	code/modules/language/undead.dm
#	code/modules/mob/living/carbon/human/examine.dm
#	code/modules/mob/living/carbon/human/species.dm
#	code/modules/mob/living/carbon/human/species_types/elf/elfd.dm
#	code/modules/mob/living/carbon/human/species_types/halfling/_halfling.dm
#	code/modules/mob/living/carbon/human/species_types/other/halfdrow.dm
#	code/modules/spells/spell_types/pointed/healing.dm
#	code/modules/spells/spell_types/pointed/ocean_embrace.dm
#	code/modules/spells/spell_types/pointed/revive.dm
#	code/modules/spells/spell_types/undirected/bardic/piper.dm
#	code/modules/spells/spell_types/undirected/bardic/seliloquy.dm
#	code/modules/spells/spell_types/undirected/conjure_item/summon_leech.dm
#	code/modules/spells/spell_types/undirected/conjure_item/summon_trident.dm
#	code/modules/spells/spell_types/undirected/list_target/convert_role/church.dm
#	code/modules/spells/spell_types/undirected/troll_shape.dm
#	icons/turf/floors.dmi
#	strings/rt/Lore_Primer.txt
#	vanderlin.dme

+++ code/modules/food_and_drinks/food/snacks.dm
@@ -40,7 +40,8 @@ All foods are distributed among various categories. Use common sense.
 	grind_results = list() //To let them be ground up to transfer their reagents
 	possible_item_intents = list(/datum/intent/food)
 	foodtype = GRAIN
-	list_reagents = list(/datum/reagent/consumable/nutriment = 1)
+	list_reagents = list()
+	var/nutrition = 1
 	w_class = WEIGHT_CLASS_SMALL
 	var/transfers_tastes = FALSE
 	var/bitesize = 3 // how many times you need to bite to consume it fully
@@ -61,6 +62,7 @@ All foods are distributed among various categories. Use common sense.
 	var/list/bonus_reagents //the amount of reagents (usually nutriment and vitamin) added to crafted/cooked snacks, on top of the ingredients reagents.
 	var/customfoodfilling = 1 // whether it can be used as filling in custom food
 	var/list/tastes  // for example list("crisps" = 2, "salt" = 1)
+	var/naturalist = FALSE
 
 	var/cooking = 0
 	var/cooktime = 25 SECONDS
@@ -93,7 +95,6 @@ All foods are distributed among various categories. Use common sense.
 	var/list/raritymod = null
 
 	var/modified = FALSE // for tracking if food has been changed
-	var/quality = 1  // used to track foodbuffs and such. Somewhat basic, could be combined with the foodbuff system directly perhaps
 
 	var/biting // if TRUE changes the icon state to the bitecount, for stuff like handpies. Will break unless you also set a base_icon_state
 	var/rot_away_timer
@@ -183,6 +184,7 @@ All foods are distributed among various categories. Use common sense.
 		add_overlay(rotflies)
 		name = "rotten [initial(name)]"
 		eat_effect = /datum/status_effect/debuff/rotfood
+		reagents.add_reagent(/datum/reagent/yuck, 5)
 		slices_num = 0
 		slice_path = null
 		cooktime = 0
@@ -215,7 +217,7 @@ All foods are distributed among various categories. Use common sense.
 
 /obj/item/reagent_containers/food/snacks/heating_act(atom/A)
 	var/obj/item/result = new /obj/item/reagent_containers/food/snacks/badrecipe(A)
-	initialize_cooked_food(result, 1)
+	initialize_cooked_food(list(result), 1)
 	return result
 
 /obj/item/proc/burning(input as num)
@@ -237,15 +239,9 @@ All foods are distributed among various categories. Use common sense.
 			burn()
 
 /obj/item/reagent_containers/food/snacks/add_initial_reagents()
-	if(!LAZYLEN(tastes))
-		return ..()
-	if(list_reagents)
-		for(var/rid in list_reagents)
-			var/amount = list_reagents[rid]
-			if(rid == /datum/reagent/consumable/nutriment || rid == /datum/reagent/consumable/nutriment/vitamin)
-				reagents.add_reagent(rid, amount, tastes.Copy())
-			else
-				reagents.add_reagent(rid, amount)
+	if(nutrition)
+		reagents.add_reagent(/datum/reagent/consumable/nutriment, nutrition, length(tastes) ? list("tastes" = tastes) : null)
+	..()
 
 /obj/item/reagent_containers/food/snacks/on_consume(mob/living/eater)
 	if(!eater)
@@ -297,6 +293,8 @@ All foods are distributed among various categories. Use common sense.
 						eater.add_stress(/datum/stress_event/noble_ate_without_table) // look i just had to okay?
 						if (prob(25))
 							to_chat(eater, span_red("I should really eat this at a table..."))
+					if(!(locate(/obj/item/plate) in range(1, eater)))
+						eater.add_stress(/datum/stress_event/noble_ate_without_plate)
 				switch (faretype)
 					if (FARE_IMPOVERISHED)
 						eater.add_stress(/datum/stress_event/noble_impoverished_food)
@@ -356,39 +354,43 @@ All foods are distributed among various categories. Use common sense.
 
 	update_appearance(UPDATE_ICON_STATE)
 
-/obj/item/reagent_containers/food/snacks/attack_self(mob/user, params)
+/obj/item/reagent_containers/food/snacks/attack_self(mob/user, list/modifiers)
 	return
 
-/obj/item/reagent_containers/food/snacks/attack(mob/living/M, mob/living/user, def_zone)
+/obj/item/reagent_containers/food/snacks/attack(mob/living/M, mob/living/user, list/modifiers)
 	if(user.used_intent.type != /datum/intent/food && (!(M == user) && isanimal(M)))
 		return ..()
 	if(!eatverb)
 		eatverb = pick("bite","chew","nibble","gnaw","gobble","chomp")
+
 	if(iscarbon(M))
 		if(!canconsume(M, user))
 			return FALSE
 
-		var/fullness = M.nutrition + 10
-		for(var/datum/reagent/consumable/C in M.reagents.reagent_list) //we add the nutrition value of what we're currently digesting
-			fullness += C.nutriment_factor * C.volume / C.metabolization_rate
+		var/obj/item/kitchen/fork/fork_check = user.get_active_held_item()
+		var/obj/item/plate/plate_check
+
+		if(istype(loc,/obj/item/plate))
+			plate_check = loc
+
+		if(fork_check)
+			if(!plate_check)
+				if(HAS_TRAIT(M,TRAIT_NOBLE_BLOOD))
+					M.add_stress(/datum/stress_event/noble_ate_with_just_a_fork)
+			else
+				if(plate_check.dirty)
+					M.add_stress(/datum/stress_event/dirty_platter)
+				else if(faretype != FARE_LAVISH && !plate_check.dirty)
+					faretype += 1
+				plate_check.fork_usages +=1
+				if(plate_check.fork_usages >= plate_check.max_fork_usages && !plate_check.dirty)
+					plate_check.dirty = TRUE
+					var/datum/component/particle_spewer = plate_check.GetComponent(/datum/component/particle_spewer/sparkle)
+					if(particle_spewer)
+						qdel(particle_spewer)
+					plate_check.update_appearance(UPDATE_OVERLAYS)
 
 		if(M == user)								//If you're eating it myself.
-/*			if(junkiness && M.satiety < -150 && M.nutrition > NUTRITION_LEVEL_STARVING + 50 && !HAS_TRAIT(user, TRAIT_VORACIOUS))
-				to_chat(M, "<span class='warning'>I don't feel like eating any more junk food at the moment!</span>")
-				return FALSE
-			else if(fullness <= 50)
-				user.visible_message("<span class='notice'>[user] hungrily [eatverb]s \the [src], gobbling it down!</span>", "<span class='notice'>I hungrily [eatverb] \the [src], gobbling it down!</span>")
-			else if(fullness > 50 && fullness < 150)
-				user.visible_message("<span class='notice'>[user] hungrily [eatverb]s \the [src].</span>", "<span class='notice'>I hungrily [eatverb] \the [src].</span>")
-			else if(fullness > 150 && fullness < 500)
-				user.visible_message("<span class='notice'>[user] [eatverb]s \the [src].</span>", "<span class='notice'>I [eatverb] \the [src].</span>")
-			else if(fullness > 500 && fullness < 600)
-				user.visible_message("<span class='notice'>[user] unwillingly [eatverb]s a bit of \the [src].</span>", "<span class='notice'>I unwillingly [eatverb] a bit of \the [src].</span>")
-			else if(fullness > (600 * (1 + M.overeatduration / 2000)))	// The more you eat - the more you can eat
-				user.visible_message("<span class='warning'>[user] cannot force any more of \the [src] to go down [user.p_their()] throat!</span>", "<span class='warning'>I cannot force any more of \the [src] to go down your throat!</span>")
-				return FALSE
-			if(HAS_TRAIT(M, TRAIT_VORACIOUS))
-				M.changeNext_move(CLICK_CD_MELEE * 0.5)*/
 			switch(M.nutrition)
 				if(NUTRITION_LEVEL_FAT to INFINITY)
 					user.visible_message("<span class='notice'>[user] forces [M.p_them()]self to eat \the [src].</span>", "<span class='notice'>I force myself to eat \the [src].</span>")
@@ -397,17 +399,8 @@ All foods are distributed among various categories. Use common sense.
 				if(0 to NUTRITION_LEVEL_STARVING)
 					user.visible_message("<span class='notice'>[user] hungrily [eatverb]s \the [src], gobbling it down!</span>", "<span class='notice'>I hungrily [eatverb] \the [src], gobbling it down!</span>")
 					M.changeNext_move(CLICK_CD_MELEE * 0.5)
-/*			if(M.energy <= 50)
-				user.visible_message("<span class='notice'>[user] hungrily [eatverb]s \the [src], gobbling it down!</span>", "<span class='notice'>I hungrily [eatverb] \the [src], gobbling it down!</span>")
-			else if(M.energy > 50 && M.energy < 500)
-				user.visible_message("<span class='notice'>[user] hungrily [eatverb]s \the [src].</span>", "<span class='notice'>I hungrily [eatverb] \the [src].</span>")
-			else if(M.energy > 500 && M.energy < 1000)
-				user.visible_message("<span class='notice'>[user] [eatverb]s \the [src].</span>", "<span class='notice'>I [eatverb] \the [src].</span>")
-			if(HAS_TRAIT(M, TRAIT_VORACIOUS))
-			M.changeNext_move(CLICK_CD_MELEE * 0.5) nom nom nom*/
 		else
 			if(!isbrain(M))		//If you're feeding it to someone else.
-//				if(fullness <= (600 * (1 + M.overeatduration / 1000)))
 				if(M.nutrition in NUTRITION_LEVEL_FAT to INFINITY)
 					M.visible_message("<span class='warning'>[user] cannot force any more of [src] down [M]'s throat!</span>", \
 										"<span class='warning'>[user] cannot force any more of [src] down your throat!</span>")
@@ -425,7 +418,6 @@ All foods are distributed among various categories. Use common sense.
 				if(!do_after(user, 3 SECONDS, M))
 					return
 				log_combat(user, M, "fed", reagents.log_list())
-//				M.visible_message("<span class='danger'>[user] forces [M] to eat [src]!</span>", "<span class='danger'>[user] forces you to eat [src]!</span>")
 			else
 				to_chat(user, "<span class='warning'>[M] doesn't seem to have a mouth!</span>")
 				return
@@ -433,7 +425,7 @@ All foods are distributed among various categories. Use common sense.
 		if(reagents)								//Handle ingestion of the reagent.
 			if(M.satiety > -200)
 				M.satiety -= junkiness
-			playsound(M.loc,'sound/misc/eat.ogg', rand(30,60), TRUE)
+			playsound(M,'sound/misc/eat.ogg', rand(30,60), TRUE)
 			if(reagents.total_volume)
 				SEND_SIGNAL(src, COMSIG_FOOD_EATEN, M, user)
 				SEND_SIGNAL(M, COMSIG_MOB_FOOD_EAT, src)
@@ -441,14 +433,21 @@ All foods are distributed among various categories. Use common sense.
 				var/amt2take = reagents.total_volume / (bitesize - bitecount)
 				if((bitecount >= bitesize) || (bitesize == 1))
 					amt2take = reagents.total_volume
+
 				reagents.trans_to(M, amt2take, transfered_by = user, method = INGEST)
+
+				if(M.has_quirk(/datum/quirk/boon/naturalist) && naturalist)
+					for(var/datum/reagent/R in reagents.reagent_list)
+						var/bonus_amount = (R.volume / reagents.total_volume) * amt2take * 0.5
+						M.reagents.add_reagent(R.type, bonus_amount)
+
 				bitecount++
 				on_consume(M)
 				checkLiked(fraction, M)
 				if(bitecount >= bitesize && !QDELETED(src))
 					qdel(src)
 				return TRUE
-		playsound(M.loc,'sound/misc/eat.ogg', rand(30,60), TRUE)
+		playsound(M,'sound/misc/eat.ogg', rand(30,60), TRUE)
 		qdel(src)
 		return FALSE
 	else if(isanimal(M))
@@ -473,7 +472,7 @@ All foods are distributed among various categories. Use common sense.
 				. += "[src] was bitten multiple times!"
 
 
-/obj/item/reagent_containers/food/snacks/attackby(obj/item/W, mob/user, params)
+/obj/item/reagent_containers/food/snacks/attackby(obj/item/W, mob/user, list/modifiers)
 	. = ..()
 	if(istype(W, /obj/item/storage))
 		..() // -> item/attackby()
@@ -485,12 +484,15 @@ All foods are distributed among various categories. Use common sense.
 		if(slice_bclass == BCLASS_CHOP)
 			user.visible_message("<span class='notice'>[user] chops [src]!</span>")
 			slice(W, user)
+			user.nobles_seen_servant_work()
 			return TRUE
 		if(slice_bclass == BCLASS_CUT)
 			user.visible_message("<span class='notice'>[user] slices [src]!</span>")
 			slice(W, user)
+			user.nobles_seen_servant_work()
 			return TRUE
 		else if(slice(W, user))
+			user.nobles_seen_servant_work()
 			return TRUE
 
 /obj/item/reagent_containers/food/snacks/proc/slice(obj/item/W, mob/user)
@@ -507,9 +509,9 @@ All foods are distributed among various categories. Use common sense.
 		return FALSE
 
 	if(slice_sound)
-		playsound(get_turf(user), 'sound/foley/slicing.ogg', 60, TRUE, -1) // added some choppy sound
+		playsound(user, 'sound/foley/slicing.ogg', 60, TRUE, -1) // added some choppy sound
 	if(chopping_sound)
-		playsound(get_turf(user), 'sound/foley/chopping_block.ogg', 60, TRUE, -1) // added some choppy sound
+		playsound(user, 'sound/foley/chopping_block.ogg', 60, TRUE, -1) // added some choppy sound
 	if(slice_batch)
 		if(!do_after(user, 3 SECONDS, src))
 			return FALSE
@@ -563,29 +565,29 @@ All foods are distributed among various categories. Use common sense.
 	add_overlay(filling)
 
 // initialize_cooked_food() is called when microwaving the food
-/obj/item/reagent_containers/food/snacks/proc/initialize_cooked_food(obj/item/reagent_containers/food/snacks/S, cooking_efficiency = 1)
-	if(reagents)
-		reagents.trans_to(S, reagents.total_volume)
-	if(S.bonus_reagents && S.bonus_reagents.len)
-		for(var/r_id in S.bonus_reagents)
-			var/amount = S.bonus_reagents[r_id] * cooking_efficiency
-			if(r_id == /datum/reagent/consumable/nutriment || r_id == /datum/reagent/consumable/nutriment/vitamin)
+/obj/item/reagent_containers/food/snacks/proc/initialize_cooked_food(list/obj/item/reagent_containers/food/snacks/outputs, cooking_efficiency = 1)
+	if(istype(outputs, /obj/item/reagent_containers/food/snacks))
+		outputs = list(outputs)
+	for(var/obj/item/reagent_containers/food/snacks/S in outputs)
+		if(reagents)
+			for(var/datum/reagent/R in reagents.reagent_list)
+				if(istype(R, /datum/reagent/consumable/nutriment))
+					continue
+				reagents.trans_id_to(S, R.type, R.volume / outputs.len)
+			for(var/r_id in bonus_reagents)
+				var/amount = bonus_reagents[r_id] * cooking_efficiency / outputs.len
 				S.reagents.add_reagent(r_id, amount)
-			else
-				S.reagents.add_reagent(r_id, amount)
-
-	if(transfers_tastes)
-		S.foodtype |= foodtype
-		S.tastes |= tastes
-
-	S.filling_color = filling_color
-	S.update_snack_overlays(src)
+		if(transfers_tastes)
+			S.foodtype |= foodtype
+			S.tastes |= tastes
+		S.filling_color = filling_color
+		S.update_snack_overlays(src)
 
 /obj/item/reagent_containers/food/snacks/proc/changefood(path, mob/living/eater)
 	if(!path || !eater)
 		return
 	var/turf/T = get_turf(eater)
-	if(eater.dropItemToGround(src))
+	if(!QDELETED(src) && eater.dropItemToGround(src))
 		qdel(src)
 	var/obj/item/I = new path(T)
 	eater.put_in_active_hand(I, ignore_animation = TRUE)
@@ -598,7 +600,7 @@ All foods are distributed among various categories. Use common sense.
 				M.emote("me", 1, "nibbles away at \the [src]")
 			bitecount++
 			SEND_SIGNAL(L, COMSIG_MOB_FEED, src, 30)
-			playsound(L.loc, 'sound/misc/eat.ogg', 25, TRUE)
+			playsound(L, 'sound/misc/eat.ogg', 25, TRUE)
 			L.taste(reagents) // why should carbons get all the fun?
 			if(bitecount >= 5)
 				var/sattisfaction_text = pick("burps from enjoyment", "meows for more", "looks at the area where \the [src] was")
@@ -606,7 +608,7 @@ All foods are distributed among various categories. Use common sense.
 					M.emote("me", 1, "[sattisfaction_text]")
 				qdel(src)
 
-/obj/item/reagent_containers/food/snacks/afterattack(obj/item/reagent_containers/M, mob/user, proximity)
+/obj/item/reagent_containers/food/snacks/afterattack(obj/item/reagent_containers/M, mob/user, proximity, list/modifiers)
 	. = ..()
 	if(!dunkable || !proximity)
 		return
@@ -628,7 +630,7 @@ All foods are distributed among various categories. Use common sense.
 	w_class = WEIGHT_CLASS_NORMAL
 	var/stored_item = 0
 
-/obj/item/reagent_containers/food/snacks/store/attackby(obj/item/W, mob/living/user, params)
+/obj/item/reagent_containers/food/snacks/store/attackby(obj/item/W, mob/living/user, list/modifiers)
 	..()
 	if(W.w_class <= WEIGHT_CLASS_SMALL & !istype(W, /obj/item/reagent_containers/food/snacks)) //can't slip snacks inside, they're used for custom foods.
 		if(W.get_sharpness())
@@ -665,7 +667,8 @@ All foods are distributed among various categories. Use common sense.
 	name = "burned mess"
 	desc = ""
 	icon_state = "badrecipe"
-	list_reagents = list(/datum/reagent/toxin/bad_food = 30)
+	nutrition = SNACK_POOR
+	list_reagents = list(/datum/reagent/toxin/bad_food = 10)
 	filling_color = "#8B4513"
 	faretype = FARE_IMPOVERISHED
 	foodtype = GROSS
